refactor(cli): dedupe result rendering helpers
